backend/pyproject.toml
[project]
name = "contabilita-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-nested-admin>=4.1.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
]
